# docs/swagger.yaml
definitions:
  makerchecker.UpdateMakerchecker:
    properties:
      id:
        type: string
      status:
        type: string
    required:
    - id
    - status
    type: object
...
  models.HttpError:
    properties:
      code:
        example: 400
        type: integer
      data:
        additionalProperties: true
        type: object
      message:
        example: status bad request
        type: string
    type: object
  models.Makerchecker:
    properties:
      _id:
        type: string
      checkerEmail:
        type: string
      checkerId:
        type: string
      data:
        additionalProperties: true
        type: object
      endpoint:
        type: string
      makerEmail:
        type: string
      makerId:
        type: string
      status:
        type: string
    required:
    - checkerId
    - data
    - endpoint
    type: object
  models.Permission:
    properties:
      _id:
        type: string
      checker:
        items:
          type: number
        type: array
      endpoint:
        type: string
      maker:
        items:
          type: number
        type: array
    type: object
  permission.UpdatePermission:
    properties:
      checker:
        items:
          type: number
        type: array
      endpoint:
        type: string
      maker:
        items:
          type: number
        type: array
    type: object
info:
  contact: {}
paths:
  /permission:
    get:
      description: Retrieves a list of makerchecker permission
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get all Makerchecker Permission
      tags:
      - permission
    post:
      description: Add a Makerchecker Permission
      parameters:
      - description: Permission Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: Invalid permission object or endpoint already exists
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Add a Makerchecker Permission
      tags:
      - permission
  /permission/{id}:
    delete:
      description: Delete a Makerchecker Permission By ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: Id cannot be empty
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No permission found with Id
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Delete a Makerchecker Permission by ID
      tags:
      - permission
    get:
      description: Retrieve a Makerchecker permission by Id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: Id cannot be empty
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No permission can be found with Id
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get Makerchecker permission by Id
      tags:
      - permission
    put:
      description: Update Makerchecker Permission by Id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/permission.UpdatePermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: Id cannot be empty and permission object is invalid
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No permission found
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Update Makerchecker Permission by Id
      tags:
      - permission
  /permission/by-endpoint:
    get:
      description: Retrieve a Makerchecker Permission by Endpoint route
      parameters:
      - description: endpoint
        in: body
        name: endpoint
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "400":
          description: Endpoint cannot be empty
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: Makerchecker permission cannot be found with endpoint route
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get Makerchecker Permission by Endpoint route
      tags:
      - permission
  /record:
    get:
      description: Retrieves a list of Makerchecker records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Makerchecker'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get all Makerchecker Record
      tags:
      - makerchecker
    post:
      description: Create a Makerchecker
      parameters:
      - description: makerchecker
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Makerchecker'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Makerchecker'
        "400":
          description: Bad request due to invalid JSON body
          schema:
            $ref: '#/definitions/models.HttpError'
        "403":
          description: Not enough permission to do makerchecker
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: Unable to find resources
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Create a Makerchecker
      tags:
      - makerchecker
    put:
      description: Update Makerchecker by approving or rejecting request
      parameters:
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/makerchecker.UpdateMakerchecker'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Makerchecker'
        "400":
          description: Bad request due to invalid JSON body
          schema:
            $ref: '#/definitions/models.HttpError'
        "403":
          description: User is not authorize to approve the request
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No makerchecker record found with makercheckerId
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Update Makerchecker by approving or rejecting request
      tags:
      - makerchecker
...
swagger: "2.0"